// github/types/paths/user/emails.types.ts
export type usersListEmailsForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: usersListEmailsForAuthenticatedUser_Query;
    querystring: never;
    path: never;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: {
          Link: { schema: string };
        };
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Array<email>;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * OAuth app tokens and personal access tokens (classic) need the `user` scope to use this endpoint.
 */
export type usersAddEmailForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: never;
    headers: never;
    cookie: never;
    body:
      | {
          /**
           * Adds one or more email addresses to your GitHub account. Must contain at least one email address. **Note:** Alternatively, you can pass a single email address or an `array` of emails addresses directly, but we recommend that you pass an object using the `emails` key.
           * @example []
           */
          emails: Array<string>;
        }
      | Array<string>
      | string;
    context: Context;
    response: ResponseBuilderFactory<{
      201: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Array<email>;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: not_found;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * OAuth app tokens and personal access tokens (classic) need the `user` scope to use this endpoint.
 */
export type usersDeleteEmailForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: never;
    headers: never;
    cookie: never;
    body:
      | {
          /**
           * Email addresses associated with the GitHub user account.
           */
          emails: Array<string>;
        }
      | Array<string>
      | string;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: not_found;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type usersListEmailsForAuthenticatedUser_Query = {
  per_page?: number;
  page?: number;
};